Wrap tree utils in createTreeUtils function to add scope
